# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/walkbg
  description: Generated server url
...
paths:
...
  /api/v1/routes/{id}:
    get:
      tags:
      - 路线管理
      summary: 查询路线详情
      description: 根据路线ID获取详细信息
      operationId: getRouteById
      parameters:
      - name: id
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: userId
        in: query
        description: 用户ID
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRouteDetailResponse'
    put:
      tags:
      - 路线管理
      summary: 更新路线
      description: 更新指定ID的路线信息
      operationId: updateRoute
      parameters:
      - name: id
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRouteBasicResponse'
    delete:
      tags:
      - 路线管理
      summary: 删除路线
      description: 删除指定ID的路线
      operationId: deleteRoute
      parameters:
      - name: id
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/v1/routes:
    get:
      tags:
      - 路线管理
      summary: 分页查询路线列表
      description: 获取路线列表，支持分页
      operationId: getRoutes
      parameters:
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: keyword
        in: query
        description: 关键词搜索
        required: false
        schema:
          type: string
      - name: regionId
        in: query
        description: 区域ID
        required: false
        schema:
          type: string
      - name: difficulty
        in: query
        description: 难度等级
        required: false
        schema:
          type: integer
          format: int32
      - name: routeType
        in: query
        description: 路线类型
        required: false
        schema:
          type: integer
          format: int32
      - name: minDistance
        in: query
        description: 最小距离
        required: false
        schema:
          type: number
          format: double
      - name: maxDistance
        in: query
        description: 最大距离
        required: false
        schema:
          type: number
          format: double
      - name: userId
        in: query
        description: 用户ID
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageRouteBasicResponse'
    post:
      tags:
      - 路线管理
      summary: 创建路线
      description: 创建新的路线
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseRouteBasicResponse'
  /api/v1/routes/{id}/favorite:
    post:
      tags:
      - 路线管理
      summary: 收藏路线
      description: 将指定路线添加到收藏
      operationId: favoriteRoute
      parameters:
      - name: id
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: userId
        in: query
        description: 用户ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
    delete:
      tags:
      - 路线管理
      summary: 取消收藏路线
      description: 从收藏中移除指定路线
      operationId: unfavoriteRoute
      parameters:
      - name: id
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: userId
        in: query
        description: 用户ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
  /api/v1/routes/{id}/complete:
    post:
      tags:
      - 路线管理
      summary: 完成路线
      description: 标记路线为已完成
      operationId: completeRoute
      parameters:
      - name: id
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: userId
        in: query
        description: 用户ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  /api/water-sources/route/{routeId}:
    get:
      tags:
      - 水源管理
      summary: 获取路线的水源
      description: 获取指定路线的所有水源
      operationId: getWaterSourcesByRoute
      parameters:
      - name: routeId
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: request
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/BaseQueryRequest'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageWaterSource'
...
  /api/v1/routes/recommendations:
    get:
      tags:
      - 路线管理
      summary: 获取推荐路线
      description: 根据用户偏好推荐路线
      operationId: getRecommendedRoutes
      parameters:
      - name: userId
        in: query
        description: 用户ID
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: type
        in: query
        description: 推荐类型
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageRouteBasicResponse'
  /api/v1/routes/nearby:
    get:
      tags:
      - 路线管理
      summary: 获取附近的路线
      description: 根据地理位置获取附近的路线
      operationId: getNearbyRoutes
      parameters:
      - name: latitude
        in: query
        description: 纬度
        required: true
        schema:
          type: number
          format: double
      - name: longitude
        in: query
        description: 经度
        required: true
        schema:
          type: number
          format: double
      - name: radius
        in: query
        description: 搜索半径（公里）
        required: false
        schema:
          type: number
          format: double
          default: 10.0
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageRouteBasicResponse'
  /api/v1/routes/my:
    get:
      tags:
      - 路线管理
      summary: 查询我创建的路线
      description: 获取当前用户创建的路线列表
      operationId: getMyRoutes
      parameters:
      - name: userId
        in: query
        description: 用户ID
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageRouteBasicResponse'
  /api/v1/routes/favorites:
    get:
      tags:
      - 路线管理
      summary: 查询我收藏的路线
      description: 获取当前用户收藏的路线列表
      operationId: getFavoriteRoutes
      parameters:
      - name: userId
        in: query
        description: 用户ID
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageRouteBasicResponse'
  /api/v1/routes/completed:
    get:
      tags:
      - 路线管理
      summary: 查询我完成的路线
      description: 获取当前用户完成的路线列表
      operationId: getCompletedRoutes
      parameters:
      - name: userId
        in: query
        description: 用户ID
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageRouteBasicResponse'
...
  /api/supplies/route/{routeId}:
    get:
      tags:
      - 补给管理
      summary: 获取路线的补给点
      description: 获取指定路线的所有补给点
      operationId: getSuppliesByRoute
      parameters:
      - name: routeId
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: request
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/BaseQueryRequest'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageSupply'
...
  /api/campsites/route/{routeId}:
    get:
      tags:
      - 营地管理
      summary: 获取路线的营地
      description: 获取指定路线的所有营地
      operationId: getCampsitesByRoute
      parameters:
      - name: routeId
        in: path
        description: 路线ID
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 页码，从0开始
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: 每页大小
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCampsite'
...
components:
  schemas:
...
    RouteCreateRequest:
      required:
      - campsites
      - created_by
      - dailyPlans
      - hitchhikeContacts
      - images
      - markerPoints
      - name
      - seasons
      - segments
      - status
      - supplies
      - tags
      - waterSources
      - waypoints
      type: object
      properties:
        id:
          type: string
        name:
          maxLength: 200
          minLength: 0
          type: string
        description:
          maxLength: 2000
          minLength: 0
          type: string
        region:
          maxLength: 100
          minLength: 0
          type: string
        region_id:
          maxLength: 64
          minLength: 0
          type: string
        distance:
          maximum: 10000.0
          exclusiveMaximum: false
          minimum: 0.0
          exclusiveMinimum: false
          type: number
        duration:
          maximum: 8760
          minimum: 0
          type: integer
          format: int32
        latitude:
          maximum: 90.0
          exclusiveMaximum: false
          minimum: -90.0
          exclusiveMinimum: false
          type: number
        longitude:
          maximum: 180.0
          exclusiveMaximum: false
          minimum: -180.0
          exclusiveMinimum: false
          type: number
        altitude:
          maximum: 10000.0
          exclusiveMaximum: false
          minimum: -500.0
          exclusiveMinimum: false
          type: number
        elevation_gain:
          maximum: 10000.0
          exclusiveMaximum: false
          minimum: 0.0
          exclusiveMinimum: false
          type: number
        elevation_loss:
          maximum: 10000.0
          exclusiveMaximum: false
          minimum: 0.0
          exclusiveMinimum: false
          type: number
        difficulty:
          maximum: 5
          minimum: 1
          type: integer
          format: int32
        route_type:
          maximum: 3
          minimum: 0
          type: integer
          format: int32
        route_direction:
          maximum: 360
          minimum: 0
          type: integer
          format: int32
        status:
          maximum: 2
          minimum: 0
          type: integer
          format: int32
        cover_url:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://.*|)$
          type: string
        map_data_id:
          maxLength: 64
          minLength: 0
          type: string
        default_map_id:
          maxLength: 64
          minLength: 0
          type: string
        created_by:
          maxLength: 64
          minLength: 0
          type: string
        tags:
          maxItems: 20
          minItems: 0
          type: array
          items:
            type: string
        seasons:
          maxItems: 4
          minItems: 0
          type: array
          items:
            type: string
        waypoints:
          maxItems: 100
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/WaypointCreateRequest'
        segments:
          maxItems: 50
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/SegmentCreateRequest'
        images:
          maxItems: 20
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/RouteImageCreateRequest'
        supplies:
          maxItems: 50
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/SupplyCreateRequest'
        campsites:
          maxItems: 30
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/CampsiteCreateRequest'
        markerPoints:
          maxItems: 100
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/MarkerPointCreateRequest'
        dailyPlans:
          maxItems: 30
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/DailyPlanCreateRequest'
        waterSources:
          maxItems: 50
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/WaterSourceCreateRequest'
        hitchhikeContacts:
          maxItems: 20
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/HitchhikeContactCreateRequest'
...
    WaypointCreateRequest:
      required:
      - latitude
      - longitude
      - name
      - sequence_number
      type: object
      properties:
        id:
          maxLength: 64
          minLength: 0
          type: string
        name:
          maxLength: 200
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        latitude:
          maximum: 90.0
          exclusiveMaximum: false
          minimum: -90.0
          exclusiveMinimum: false
          type: number
          format: double
        longitude:
          maximum: 180.0
          exclusiveMaximum: false
          minimum: -180.0
          exclusiveMinimum: false
          type: number
          format: double
        elevation:
          maximum: 10000.0
          exclusiveMaximum: false
          minimum: -500.0
          exclusiveMinimum: false
          type: number
          format: double
        type:
          maxLength: 50
          minLength: 0
          type: string
        icon_url:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://.*|)$
          type: string
        image_url:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://.*|)$
          type: string
        sequence_number:
          maximum: 1000
          minimum: 1
          type: integer
          format: int32
    ApiResponseRouteBasicResponse:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/RouteBasicResponse'
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
...
    ApiResponseString:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: string
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
...
    BaseQueryRequest:
      required:
      - page
      - size
      - sortBy
      - sortDir
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
        sortBy:
          type: string
        sortDir:
          type: string
    ApiResponsePageWaterSource:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageWaterSource'
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
...
    ApiResponsePageRouteBasicResponse:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageRouteBasicResponse'
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
...
    ApiResponseRouteDetailResponse:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/RouteDetailResponse'
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
...
    RouteDetailResponse:
      required:
      - campsites
      - created_at
      - daily_plans
      - hitchhike_contacts
      - id
      - image_urls
      - is_favorite
      - is_loop
      - marker_points
      - name
      - popularity
      - segments
      - status
      - supplies
      - tags
      - updated_at
      - usage_count
      - water_sources
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        region_id:
          type: string
        region:
          type: string
        distance:
          type: number
          format: double
        duration:
          type: integer
          format: int32
        elevation_gain:
          type: number
          format: double
        elevation_loss:
          type: number
          format: double
        difficulty:
          type: integer
          format: int32
        route_type:
          type: integer
          format: int32
        route_direction:
          type: integer
          format: int32
        cover_url:
          type: string
        default_map_id:
          type: string
        popularity:
          type: integer
          format: int32
        usage_count:
          type: integer
          format: int32
        is_loop:
          type: boolean
        is_favorite:
          type: boolean
        status:
          type: integer
          format: int32
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
        created_by:
          type: string
        creator:
          $ref: '#/components/schemas/UserBasicDto'
        segments:
          type: array
          items:
            $ref: '#/components/schemas/SegmentDto'
        daily_plans:
          type: array
          items:
            $ref: '#/components/schemas/DailyPlanDto'
        tags:
          type: array
          items:
            type: string
        image_urls:
          type: array
          items:
            type: string
        ratings:
          $ref: '#/components/schemas/RatingDto'
        weather_info:
          $ref: '#/components/schemas/WeatherInfoDto'
        water_sources:
          type: array
          items:
            $ref: '#/components/schemas/WaterSourceDto'
        campsites:
          type: array
          items:
            $ref: '#/components/schemas/CampsiteDto'
        supplies:
          type: array
          items:
            $ref: '#/components/schemas/SupplyDto'
        hitchhike_contacts:
          type: array
          items:
            $ref: '#/components/schemas/HitchhikeContactDto'
        marker_points:
          type: array
          items:
            $ref: '#/components/schemas/MarkerPointDto'
...
    ApiResponsePageSupply:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PageSupply'
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
...
    PageCampsite:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Campsite'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...